reduce amount of sync tasks by a factor of 4
